Fixed precision bug in Fid class

+++ fid.cpp
@@ -102,7 +102,7 @@ int Fid::size() const
 
 double Fid::at(const int i) const
 {
-    return (double)atNorm(i)*data->vMult;
+    return atNorm(i)*data->vMult;
 }
 
 double Fid::spacing() const
@@ -147,12 +147,12 @@ qint64 Fid::atRaw(const int i) const
     return data->fid.at(i);
 }
 
-qint64 Fid::atNorm(const int i) const
+double Fid::atNorm(const int i) const
 {
     if(data->shots > 1)
-        return data->fid.at(i)/data->shots;
+        return (double)data->fid.at(i)/(double)data->shots;
     else
-        return data->fid.at(i);
+        return (double)data->fid.at(i);
 }
 
 qint64 Fid::shots() const
